Revert "Auth feature (OCR)"
