+++ App.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import { StatusBar, StyleSheet, useColorScheme, View } from 'react-native';
-// import FireAuth from './src/pages/fireauth';
-import OCR from './src/ocr/ocr'; 
+import FireAuth from './src/pages/fireauth';
 
 const App = () => {
   const isDarkMode = useColorScheme() === 'dark';
@@ -9,7 +8,7 @@ const App = () => {
   return (
     <View style={styles.container}>
       <StatusBar barStyle={isDarkMode ? 'light-content' : 'dark-content'} />
-      <OCR />
+      <FireAuth />
     </View>
   );
 };
