+++ android/app/build.gradle
@@ -1,9 +1,6 @@
 apply plugin: "com.android.application"
 apply plugin: "org.jetbrains.kotlin.android"
 apply plugin: "com.facebook.react"
-apply plugin: 'com.google.gms.google-services'
-
-
 
 /**
  * This is the configuration block to customize your React Native Android app.
@@ -80,14 +77,13 @@ android {
     buildToolsVersion rootProject.ext.buildToolsVersion
     compileSdk rootProject.ext.compileSdkVersion
 
-    namespace "com.expensely"
+    namespace "com.firebase_auth"
     defaultConfig {
-        applicationId "com.expensely"
+        applicationId "com.firebase_auth"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
         versionName "1.0"
-        multiDexEnabled true
     }
     signingConfigs {
         debug {
@@ -120,8 +116,4 @@ dependencies {
     } else {
         implementation jscFlavor
     }
-    implementation 'com.google.firebase:firebase-auth:22.3.0'
-    implementation 'com.google.android.gms:play-services-auth:20.7.0'
 }
-
-apply from: "../../node_modules/react-native-vector-icons/fonts.gradle"
