+++ package.json
@@ -10,20 +10,9 @@
     "test": "jest"
   },
   "dependencies": {
-    "@react-native-async-storage/async-storage": "^1.24.0",
-    "@react-native-firebase/app": "^22.2.1",
-    "@react-native-firebase/auth": "^22.2.1",
-    "@react-native-google-signin/google-signin": "^14.0.1",
-    "@react-native/new-app-screen": "0.80.0",
-    "@react-navigation/native": "^7.1.14",
-    "@react-navigation/native-stack": "^7.3.19",
     "react": "19.1.0",
     "react-native": "0.80.0",
-    "react-native-fs": "^2.20.0",
-    "react-native-image-picker": "^8.2.1",
-    "react-native-safe-area-context": "^5.5.0",
-    "react-native-screens": "^4.11.1",
-    "react-native-vector-icons": "^10.2.0"
+    "@react-native/new-app-screen": "0.80.0"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
@@ -37,8 +26,7 @@
     "@react-native/metro-config": "0.80.0",
     "@react-native/typescript-config": "0.80.0",
     "@types/jest": "^29.5.13",
-    "@types/react": "^19.1.8",
-    "@types/react-native": "^0.72.8",
+    "@types/react": "^19.1.0",
     "@types/react-test-renderer": "^19.1.0",
     "eslint": "^8.19.0",
     "jest": "^29.6.3",
