--- types/react-native-vector-icons.d.ts
@@ -1,4 +1,0 @@
-declare module 'react-native-vector-icons/MaterialIcons' {
-import { Icon } from 'react-native-vector-icons/Icon';
-export default Icon;
-}
